Add credential checking for sensitive actions using accessToken

+++ SkyForceWebService/src/main/java/com/skyforce/SkyForceWebService/controller/VendorController.java
@@ -9,10 +9,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.security.NoSuchAlgorithmException;
-import java.util.concurrent.atomic.AtomicLong;
-
 import javax.validation.Valid;
+import java.security.NoSuchAlgorithmException;
 import java.util.List;
 
 @RestController
@@ -20,53 +18,47 @@ public class VendorController {
 
     // TODO:  Change http status for different exceptions
 
-    // TODO: add service for vendor to add shop
-
     @Autowired
     VendorService vendorService;
 
     @Autowired
     ShopService shopService;
 
-    private AtomicLong nextId = new AtomicLong();
-
-    // TODO: Need a machanism to remember the last assigned id
-
-    private long getRecentId () {
-        List<Vendor> vendors = vendorService.findAll();
-        Long recentId = 0L;
-        for (Vendor vendor : vendors) {
-            if (vendor.getId() > recentId) {
-                recentId = vendor.getId();
-            }
-        }
-        return recentId;
-    }
-
-    // Get all the vendors
     @GetMapping("/vendors")
     public String getAllVendors() {
         List<Vendor> vendors = vendorService.findAll();
         return JSONConvert.JSONConverter(vendors);
     }
+    // pass
 
-    // Get vendor by ID
+
+
     @GetMapping("/vendor")
     public String getVendorById(
-            @RequestParam Long vendorId
+            @RequestParam Long vendorId,
+            @RequestHeader(value = "Authorization") String accessToken
     ) {
-        Vendor vendor = vendorService.findVendorById(vendorId);
-        return JSONConvert.JSONConverter(vendor);
+        String[] info = ValidationController.decryptAccessToken(accessToken);
+        if (info.length != 2)
+            throw new IllegalArgumentException();
+        if (Long.parseLong(info[0]) == vendorId && info[1].equals("VENDOR")) {
+            Vendor vendor = vendorService.findVendorById(vendorId);
+            return JSONConvert.JSONConverter(vendor);
+        } else {
+            throw new IllegalArgumentException();
+        }
     }
 
     @GetMapping("/vendor/signin")
-    public String signInVendor (
+    public String signInVendor(
             @RequestParam String email,
             @RequestParam String password
     ) throws NoSuchAlgorithmException {
         Vendor vendor = vendorService.findVendorByEmail(email);
         return ValidationController.UserSignIn(vendor, password);
     }
+    // pass
+
     @ResponseStatus(value = HttpStatus.UNAUTHORIZED,
             reason = "Email or password incorrect")
     @ExceptionHandler(IllegalArgumentException.class)
@@ -85,24 +77,45 @@ public class VendorController {
         return ValidationController.getAccessToken(vendor.getId(), "VENDOR");
     }
 
+
     @PutMapping("/vendor")
     public String updateVendorById(
             @RequestParam Long vendorId,
-            @Valid @RequestBody Vendor vendor) {
-        Vendor oldVendor = vendorService.findVendorById(vendorId);
-        oldVendor.setFirstName(vendor.getFirstName());
-        oldVendor.setLastName(vendor.getLastName());
-        oldVendor.setGender(vendor.getGender());
-        Vendor updatedVendor = vendorService.save(oldVendor);
-        return JSONConvert.JSONConverter(updatedVendor) ;
+            @RequestHeader(value = "Authorization") String accessToken,
+            @Valid @RequestBody Vendor vendor
+    ) {
+        String[] info = ValidationController.decryptAccessToken(accessToken);
+        if (info.length != 2)
+            throw new IllegalArgumentException();
+        if (Long.parseLong(info[0]) == vendorId && info[1].equals("VENDOR")) {
+            Vendor oldVendor = vendorService.findVendorById(vendorId);
+            oldVendor.setFirstName(vendor.getFirstName());
+            oldVendor.setLastName(vendor.getLastName());
+            oldVendor.setGender(vendor.getGender());
+            Vendor updatedVendor = vendorService.save(oldVendor);
+            return JSONConvert.JSONConverter(updatedVendor);
+        } else {
+            throw new IllegalArgumentException();
+        }
     }
+    // pass
 
     @DeleteMapping("/vendor")
     public ResponseEntity<?> deleteVendor(
-            @RequestParam Long vendorId
+            @RequestParam Long vendorId,
+            @RequestHeader(value = "Authorization") String accessToken
     ) {
-        Vendor vendor = vendorService.findVendorById(vendorId);
-        vendorService.delete(vendor);
-        return ResponseEntity.ok().build();
+        String[] info = ValidationController.decryptAccessToken(accessToken);
+        if (info.length != 2)
+            throw new IllegalArgumentException();
+        if (Long.parseLong(info[0]) == vendorId && info[1].equals("VENDOR")) {
+            Vendor vendor = vendorService.findVendorById(vendorId);
+            vendorService.delete(vendor);
+            return ResponseEntity.ok().build();
+        } else {
+            throw new IllegalArgumentException();
+        }
     }
+    // one null pointer exp when wrong access token was given
+    // partially passed
 }
